+++ src/vocutouts/workers.py
@@ -16,7 +16,7 @@ from __future__ import annotations
 
 import os
 from datetime import datetime, timezone
-from typing import Any, Dict, List
+from typing import Any, Dict, List, Tuple
 from urllib.parse import urlparse
 from uuid import UUID
 
@@ -36,6 +36,9 @@ from lsst.image_cutout_backend.stencils import (
     SkyStencil,
 )
 
+BACKENDS: Dict[str, ImageCutoutBackend] = {}
+"""Cache of image cutout backends by Butler repository label."""
+
 redis_host = os.environ["CUTOUT_REDIS_HOST"]
 redis_password = os.getenv("CUTOUT_REDIS_PASSWORD")
 broker = RedisBroker(host=redis_host, password=redis_password)
@@ -43,13 +46,6 @@ results = RedisBackend(host=redis_host, password=redis_password)
 dramatiq.set_broker(broker)
 broker.add_middleware(CurrentMessage())
 broker.add_middleware(Results(backend=results))
-
-repository = os.environ["CUTOUT_BUTLER_REPOSITORY"]
-output_root = os.environ["CUTOUT_STORAGE_URL"]
-tmpdir = os.environ.get("CUTOUT_TMPDIR", "/tmp")
-butler = Butler(repository)
-projection_finder = projection_finders.ProjectionFinder.make_default()
-backend = ImageCutoutBackend(butler, projection_finder, output_root, tmpdir)
 
 
 # Stubs for other actors implemented elsewhere that workers may call.
@@ -81,6 +77,51 @@ class TaskFatalError(Exception):
 
 class TaskTransientError(Exception):
     """Some transient problem occurred."""
+
+
+def get_backend(butler_label: str) -> ImageCutoutBackend:
+    """Given the Butler label, retrieve or build a backend.
+
+    The dataset ID will be a URI of the form ``butler://<label>/<uuid>``.
+    Each label corresponds to a different Butler and thus a different backend,
+    which are cached in `BACKENDS`.
+
+    Parameters
+    ----------
+    butler_label : `str`
+        The label portion fo the Butler URI.
+
+    Returns
+    -------
+    backend : `lsst.image_cutout_backend.ImageCutoutBackend`
+        Backend to use.
+    """
+    if butler_label in BACKENDS:
+        return BACKENDS[butler_label]
+    butler = Butler(butler_label)
+    projection_finder = projection_finders.ProjectionFinder.make_default()
+    output = os.environ["CUTOUT_STORAGE_URL"]
+    tmpdir = os.environ.get("CUTOUT_TMPDIR", "/tmp")
+    backend = ImageCutoutBackend(butler, projection_finder, output, tmpdir)
+    BACKENDS[butler_label] = backend
+    return backend
+
+
+def parse_uri(uri: str) -> Tuple[str, UUID]:
+    """Parse a Butler URI.
+
+    Parameters
+    ----------
+    uri : `str`
+        URI to a Butler object.
+
+    Returns
+    -------
+    data : Tuple[`str`, `uuid.UUID`]
+        The Butler label and the object UUID.
+    """
+    parsed_uri = urlparse(uri)
+    return parsed_uri.netloc, UUID(parsed_uri.path[1:])
 
 
 @dramatiq.actor(queue_name="cutout", max_retries=1, store_results=True)
@@ -137,6 +178,10 @@ def cutout(
         logger.warning(msg)
         raise TaskFatalError(f"UsageError {msg}")
 
+    # Parse the dataset ID and retrieve an appropriate backend.
+    butler_label, uuid = parse_uri(dataset_ids[0])
+    backend = get_backend(butler_label)
+
     # Convert the stencils to SkyStencils.
     sky_stencils: List[SkyStencil] = []
     for stencil_dict in stencils:
@@ -161,7 +206,7 @@ def cutout(
 
     # Perform the cutout.
     try:
-        result = backend.process_uuid(sky_stencils[0], UUID(dataset_ids[0]))
+        result = backend.process_uuid(sky_stencils[0], uuid)
     except Exception as e:
         logger.exception("Cutout processing failed")
         msg = f"Error Cutout processing failed\n{type(e).__name__}: {str(e)}"